libopkg: make error messages easier to grep for
opkg is used e.g. as part of the yocto rootfs creation, where many packages are
installed. When there is an error, a log file with thousands of lines is
created, and people typically search for "error".

Thus, let's prefix error messages with "error: " instead of "* " which is easy to miss.

Before:
    * check_data_file_clashes: Package systemd wants to install file build/tmp-glibc/work/my-arch/my-image/1234/rootfs/usr/bin/foobar
    But that file is already provided by package  * foo-package

After:
    error: check_data_file_clashes: Package systemd wants to install file build/tmp-glibc/work/my-arch/my-image/1234/rootfs/usr/bin/foobar
    But that file is already provided by package foo-package

Signed-off-by: Etienne Cordonnier <ecordonnier@snap.com>
Signed-off-by: Alex Stewart <alex.stewart@ni.com>