libopkg: Improve error messages when downloading files via libcurl
So far, error messages included whatever curl_easy_strerror() reported - which
usually is something pretty generic like 'HTTP response code said error', since
it is based on the return code only. Improve upon this by using the more
specific information from the 'error buffer' (see [0]) instead, as long as it is
available.

Please note that the wording / formatting of the opkg-specific prefix in the
error messages has changed slightly as part of this change.

[0]: https://curl.se/libcurl/c/CURLOPT_ERRORBUFFER.html

Signed-off-by: Moritz Haase <Moritz.Haase@bmw.de>
Signed-off-by: Alex Stewart <alex.stewart@ni.com>