libopkg: Enable verbose libcurl output on high verbosity levels
Tell libcurl via CURLOPT_VERBOSE (see [0]) to provide verbose debugging output
on stderr if opkg's debug level is DEBUG (=3) or higher. This is useful when
investigating download issues.

Please note that the debug output might contain sensitive information - keep
that in mind when sharing logs or enabling debug logging in CI for example.

[0]: https://curl.se/libcurl/c/CURLOPT_VERBOSE.html

Signed-off-by: Moritz Haase <Moritz.Haase@bmw.de>
Signed-off-by: Alex Stewart <alex.stewart@ni.com>