parse_util: Avoid field confusion in is_field
When is_field is called and passed `type` doesn't end with a colon, it
is possible for opkg to confuse fields sharing prefix that matches
`type`. For example one could have a custom field in a `control` file
with name `Package_name`, which when present after the actual `Package`
field would overwrite and in turn produce invalid state in opkg files.

Avoid such confusion by checking for a colon at the end of either `type`
or in `line` just after `type` ends.

Signed-off-by: Michal Sieron <michalwsieron@gmail.com>
Signed-off-by: Alex Stewart <alex.stewart@emerson.com>