CONTRIBUTING: update build deps debian distro
The previous version of the build instructions implied support for
Debian 11 - which was true. Debian 12 is now the stable release, and is
what we have tested the new cmake build logic on.

Update the debian version in the instructions to 12.

Incidental. Fixup some styling in the same instructions set to be more
clear.

Signed-off-by: Alex Stewart <alex.stewart@emerson.com>