CONTRIBUTING: add simple git send-email example
While the configuration using diff.renames=copy is not strictly identical to git send-email "-M" option,
it is good enough for almost all cases and a simple one-liner.

Also, posting patches via mailing-list is rare nowadays, and not everybody knows that you need to be a subscriber
in order to be able to post a patch.

Signed-off-by: Etienne Cordonnier <ecordonnier@snap.com>
Signed-off-by: Alex Stewart <alex.stewart@emerson.com>