docker: run container as local user
Running the opkg-dev container as root tends to create permissions
issues when building distribution content.

Run the container as the local user, to match ownership information
between the host and container.

Move the container workspace mount point to /usr/local/src, since that
is a more standard build location for foreign project sources.

Signed-off-by: Alex Stewart <alex.stewart@emerson.com>